requirements.txt
beautifulsoup4==4.12.2
chromadb==0.4.12
Flask==3.0.0
gunicorn==21.2.0
langchain==0.0.298
nest-asyncio==1.5.8
openai==0.28.0
pypdf==4.0.1
PyJWT==2.8.0
requests==2.31.0
tiktoken==0.5.1
trafilatura==1.6.2
twilio==8.10.3
